refactor: restructure lib
