+++ src/components/base/index.ts
@@ -1,11 +1,8 @@
 export { Topbar } from './Topbar';
-export { Post } from './Post';
 export { TextWithHighlights } from './TextWithHighlights';
 export { FilterByChips } from './FilterByChips';
 export { FiltersMenu } from './FiltersMenu';
 export { FiltersSection } from './FiltersSection';
 export { TimeFilter } from './TimeFilter';
-export { SinglePostView } from './SinglePostView';
-export { DescriptionAccordion } from './DescriptionAccordion';
 export { Button } from './Button';
 export { MediaViewer } from './MediaViewer';
