+++ src/components/custom/DescriptionAccordion/DescriptionAccordion.style.tsx
