+++ src/components/custom/DescriptionAccordion/index.ts
