+++ src/components/custom/Post/index.ts
