+++ src/components/custom/SinglePostView/index.ts
