+++ src/components/custom/index.ts
@@ -1,0 +1,3 @@
+export { DescriptionAccordion } from './DescriptionAccordion';
+export { SinglePostView } from './SinglePostView';
+export { Post } from './Post';
